# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MyloWare API",
    "description": "Llama Stack-native multi-agent video production platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/admin/dlq": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List Dead Letters",
        "description": "List dead letter queue entries.\n\nArgs:\n    source: Optional filter by source (\"sora\", \"remotion\")\n    unresolved_only: If True, only return unresolved entries\n\nReturns:\n    List of dead letter entries",
        "operationId": "list_dead_letters_admin_dlq_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "name": "unresolved_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Unresolved Only"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeadLetterListResponse"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/dlq/{dead_letter_id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get Dead Letter",
        "description": "Get a specific dead letter entry.\n\nArgs:\n    dead_letter_id: ID of the dead letter entry\n\nReturns:\n    Dead letter details including full payload",
        "operationId": "get_dead_letter_admin_dlq__dead_letter_id__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "dead_letter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dead Letter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeadLetterDetailResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/dlq/{dead_letter_id}/replay": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Replay Dead Letter",
        "description": "Replay a dead letter entry by re-processing the webhook.\n\nThis will attempt to process the stored webhook payload again,\nwhich may trigger the workflow continuation if the underlying issue is resolved.\n\nArgs:\n    dead_letter_id: ID of the dead letter entry to replay\n\nReturns:\n    Status of replay operation",
        "operationId": "replay_dead_letter_admin_dlq__dead_letter_id__replay_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "dead_letter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dead Letter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Replay Dead Letter Admin Dlq  Dead Letter Id  Replay Post"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/dlq/{dead_letter_id}/resolve": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Resolve Dead Letter",
        "description": "Mark a dead letter as resolved.\n\nThis is useful when a dead letter has been manually handled\nor is no longer relevant.\n\nArgs:\n    dead_letter_id: ID of the dead letter entry to resolve\n\nReturns:\n    Status of resolution",
        "operationId": "resolve_dead_letter_admin_dlq__dead_letter_id__resolve_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "dead_letter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dead Letter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Resolve Dead Letter Admin Dlq  Dead Letter Id  Resolve Post"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/kb/reload": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Reload Knowledge Base",
        "description": "Reload knowledge base on-demand (honors unchanged manifest).",
        "operationId": "reload_knowledge_base_admin_kb_reload_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Reload Knowledge Base Admin Kb Reload Post"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeadLetterDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "last_attempt_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Attempt At"
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source",
          "run_id",
          "payload",
          "error",
          "attempts",
          "created_at",
          "last_attempt_at",
          "resolved_at"
        ],
        "title": "DeadLetterDetailResponse",
        "description": "Response for dead letter detail."
      },
      "DeadLetterListItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "last_attempt_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Attempt At"
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source",
          "run_id",
          "error",
          "attempts",
          "created_at",
          "last_attempt_at",
          "resolved_at"
        ],
        "title": "DeadLetterListItem",
        "description": "Dead letter queue entry for list view."
      },
      "DeadLetterListResponse": {
        "properties": {
          "dead_letters": {
            "items": {
              "$ref": "#/components/schemas/DeadLetterListItem"
            },
            "type": "array",
            "title": "Dead Letters"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "dead_letters",
          "count"
        ],
        "title": "DeadLetterListResponse",
        "description": "Response for dead letter list."
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "description": "Machine-readable error code"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail",
            "description": "Human-readable or structured error detail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}